refactor: move message and request handling to new module
